requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
Markdown==3.6
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2023.3
